requirements.txt
aiogram==2.21
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Django==4.1
django-filter==22.1
djangorestframework==3.13.1
frozenlist==1.3.1
h11==0.13.0
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.2
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
unicorn==2.0.0
uvicorn==0.18.2
yarl==1.8.1
